move the collections folder structure
